Move registration tests to be a new folder
