+++ tests/envs/registration/test_spec.py
